pyproject.toml
[project]
name = "sycore"
version = "0.1.0"
description = "Generic agent/pipeline execution framework - AI OS driver core"
readme = "README.md"
authors = [
    { name = "Sri Viswanath", email = "sridatta@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "anthropic>=0.30",
    "python-frontmatter>=1.0.0",
    "jinja2>=3.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = false
warn_unused_ignores = true
ignore_missing_imports = true
